+++ src/renderer/App.tsx
@@ -495,7 +495,7 @@ const Hello = () => {
                 display: 'flex',
                 ml: `${drawerWidth}px`,
                 boxSizing: 'border-box',
-                background: '#2E3B55', // original #2E3B55 #102a43 #486581
+                background: '#2E3B55', // original #2E3B55 #102a43 #486581 background #101218 higlight buttons #1f2d41
               }}
             >
               <Toolbar