+++ src/renderer/components/mediaAdvanced/entryInput/onList/EntryInputOnList.tsx
@@ -1,4 +1,4 @@
-import { CardContent } from '@mui/material';
+import { Card, CardContent } from '@mui/material';
 import { useAdvancedInput } from 'renderer/context/advanced/AdvancedInputContext';
 import { useAdvancedMedia } from 'renderer/context/advanced/AdvancedMediaContext';
 import MyDatePickerCompletedAt from './MyDatePickerCompletedAt';
@@ -17,33 +17,40 @@ export default function EntryInputOnList() {
   const myAdvancedInput: any = useAdvancedInput();
 
   return (
-    <CardContent
+    <Card
+      elevation={0}
       sx={{
         gridColumn: '2/3',
         gridRow: '3/6',
         border: '1px solid SteelBlue',
-        width: '100%',
-        height: '100%',
-        display: 'grid',
-        gridTemplateColumns: 'repeat(3, 1fr)', // 175 px
-        gridTemplateRows: 'repeat(4, 1fr)', // 73 px
-        padding: '16px',
-        '&:last-child': { pb: '16px' },
       }}
+      variant="outlined"
     >
-      <ProgressInput />
-      {myAdvancedMedia.advancedMedia.type === 'MANGA' ? (
-        <ProgressVolumesInput />
-      ) : null}
-      <RepeatInput />
-      <StatusSelect />
-      <ScoreInput />
-      <PrivateCheckBox />
-      <MyDatePickerStartedAt />
-      <MyDatePickerCompletedAt />
-      <NotesInput />
-      <UpdateButton />
-    </CardContent>
+      <CardContent
+        sx={{
+          display: 'grid',
+          gridTemplateColumns: 'repeat(3, 1fr)', // 175 px
+          gridTemplateRows: 'repeat(4, 1fr)', // 73 px
+          width: '100%',
+          height: '100%',
+          padding: '16px',
+          '&:last-child': { pb: '16px' },
+        }}
+      >
+        <ProgressInput />
+        {myAdvancedMedia.advancedMedia.type === 'MANGA' ? (
+          <ProgressVolumesInput />
+        ) : null}
+        <RepeatInput />
+        <StatusSelect />
+        <ScoreInput />
+        <PrivateCheckBox />
+        <MyDatePickerStartedAt />
+        <MyDatePickerCompletedAt />
+        <NotesInput />
+        <UpdateButton />
+      </CardContent>
+    </Card>
   );
 }