+++ src/renderer/components/mediaAdvanced/image/AdvancedImage.tsx
@@ -1,5 +1,5 @@
 import { Tooltip } from '@mui/joy';
-import { Box, CardActionArea } from '@mui/material';
+import { Box, Card, CardActionArea } from '@mui/material';
 import { useTitle } from 'renderer/context/TitleContext';
 import { useAdvancedDefaultLink } from 'renderer/context/advanced/AdvancedDefaultLinkContext';
 import { useAdvancedMedia } from 'renderer/context/advanced/AdvancedMediaContext';
@@ -11,56 +11,69 @@ export default function AdvancedImage() {
   const titlePreference: any = useTitle();
 
   return (
-    <Tooltip
+    <Card
+      sx={{
+        width: '250px',
+        height: '340px',
+        gridColumn: '1/2',
+        gridRow: '1/5',
+        border: '1px solid SteelBlue',
+        overflow: 'hidden',
+        display: 'flex',
+        justifyContent: 'center',
+        alignItems: 'center',
+      }}
       variant="outlined"
-      color="primary"
-      title={
-        myAdvancedDefaultLink.advancedDefaultLink === 'AniList'
-          ? 'Go to AniList'
-          : 'Go to MyAnimeList'
-      }
-      followCursor
-      placement="right-end"
     >
-      <CardActionArea
-        sx={{
-          width: '250px',
-          height: '340px',
-          gridColumn: '1/2',
-          gridRow: '1/5',
-          border: '1px solid SteelBlue',
-          overflow: 'hidden',
-        }}
-        onClick={() => {
-          console.log(titlePreference);
-          if (
-            myAdvancedMedia.advancedMedia.idMal !== null &&
-            myAdvancedDefaultLink.advancedDefaultLink === 'MyAnimeList'
-          ) {
-            window.electron.ipcRenderer.sendMessage('openExternalLink', [
-              getMalLink(
-                myAdvancedMedia.advancedMedia.idMal,
-                myAdvancedMedia.advancedMedia.type,
-              ),
-            ]);
-          } else {
-            window.electron.ipcRenderer.sendMessage('openExternalLink', [
-              myAdvancedMedia.advancedMedia.siteUrl,
-            ]);
-          }
-        }}
+      <Tooltip
+        variant="outlined"
+        color="primary"
+        title={
+          myAdvancedDefaultLink.advancedDefaultLink === 'AniList'
+            ? 'Go to AniList'
+            : 'Go to MyAnimeList'
+        }
+        followCursor
+        placement="right-end"
       >
-        <Box
-          component="img"
-          alt="null"
-          src={myAdvancedMedia.advancedMedia.image}
+        <CardActionArea
           sx={{
-            marginLeft: 'auto',
-            marginRight: 'auto',
-            display: 'block',
+            width: '250px',
+            height: '340px',
+            overflow: 'hidden',
+            display: 'flex',
           }}
-        />
-      </CardActionArea>
-    </Tooltip>
+          onClick={() => {
+            console.log(titlePreference);
+            if (
+              myAdvancedMedia.advancedMedia.idMal !== null &&
+              myAdvancedDefaultLink.advancedDefaultLink === 'MyAnimeList'
+            ) {
+              window.electron.ipcRenderer.sendMessage('openExternalLink', [
+                getMalLink(
+                  myAdvancedMedia.advancedMedia.idMal,
+                  myAdvancedMedia.advancedMedia.type,
+                ),
+              ]);
+            } else {
+              window.electron.ipcRenderer.sendMessage('openExternalLink', [
+                myAdvancedMedia.advancedMedia.siteUrl,
+              ]);
+            }
+          }}
+        >
+          <Box
+            component="img"
+            alt="null"
+            src={myAdvancedMedia.advancedMedia.image}
+            sx={{
+              marginLeft: 'auto',
+              marginRight: 'auto',
+              display: 'block',
+            }}
+          />
+        </CardActionArea>
+      </Tooltip>
+    </Card>
   );
 }