+++ src/renderer/components/mediaAdvanced/title/AdvancedTitle.tsx
@@ -1,5 +1,5 @@
 import { Tooltip } from '@mui/joy';
-import { CardContent, Typography } from '@mui/material';
+import { Card, CardContent, Typography } from '@mui/material';
 import { useEffect, useRef, useState } from 'react';
 import { useTitle } from 'renderer/context/TitleContext';
 import { useAdvancedMedia } from 'renderer/context/advanced/AdvancedMediaContext';
@@ -23,25 +23,29 @@ export default function AdvancedTitle() {
   }, [textElementRef, myAdvancedMedia]);
 
   return (
-    <CardContent
+    <Card
+      elevation={0}
       sx={{
         gridColumn: '2/3',
         gridRow: '1/2',
         border: '1px solid SteelBlue',
       }}
+      variant="outlined"
     >
-      <Tooltip
-        title={getTitle(titlePreference.title, myAdvancedMedia.advancedMedia)}
-        followCursor
-        disableHoverListener={!isOverflowed}
-        variant="outlined"
-        color="primary"
-        size="sm"
-      >
-        <Typography noWrap fontWeight="bold" ref={textElementRef}>
-          {getTitle(titlePreference.title, myAdvancedMedia.advancedMedia)}
-        </Typography>
-      </Tooltip>
-    </CardContent>
+      <CardContent>
+        <Tooltip
+          title={getTitle(titlePreference.title, myAdvancedMedia.advancedMedia)}
+          followCursor
+          disableHoverListener={!isOverflowed}
+          variant="outlined"
+          color="primary"
+          size="sm"
+        >
+          <Typography noWrap fontWeight="bold" ref={textElementRef}>
+            {getTitle(titlePreference.title, myAdvancedMedia.advancedMedia)}
+          </Typography>
+        </Tooltip>
+      </CardContent>
+    </Card>
   );
 }