+++ src/renderer/components/settings/etc/CardContentPaddingAdjusted.tsx
@@ -1,0 +1,9 @@
+import { CardContent, styled } from '@mui/material';
+
+const CardContentPaddingAdjusted = styled(CardContent)(`
+  &:last-child {
+    padding-bottom: 10px;
+  }
+`);
+
+export default CardContentPaddingAdjusted;